Moved spare LED from caps-lock to @ (OSI shift LOCK)
This allows the keymap to determine which LED is used, rather than requiring the
decision to be hardwired in at the time of assembly.  For the OSI keyboard, the
LED could be placed on the SHIFTLOCK key, at the far right LED, or in the
CAPSLOCK position, which is just to the left of the RETURN and not otherwise
populated in the OSI layout.